+++ packages/dartvel_core/test/fixtures/workers_web/worker.dart
@@ -1,0 +1,8 @@
+import 'package:dartvel_core/web_worker.dart';
+
+import 'tasks.dart';
+
+void main() {
+  registerAll();
+  dvWebWorkerMain();
+}